% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khb.R
\name{khb}
\alias{khb}
\title{Karlson-Holm-Breen method for comparing probit coefficients}
\usage{
khb(X, y, z)
}
\arguments{
\item{X}{data frame comprising independent variables including confounding variable.}

\item{y}{vector of dependent variable.}

\item{z}{character string giving the name of the confounding variable in \code{X}.}
}
\value{
\code{khb} returns for all model coefficients the p-value for the null hypothesis that the change in coefficients is not attributable to confounding by z.
}
\description{
Significance test for confounding; that is, the difference between regression 
coefficients from same-sample nested logit and probit models. The test procedure follows
Karlson et al (2012), Section 3.4.
}
\examples{
## 1. load results from Klein (2015a)
 data(klein15a)

## 2. apply KHB method
with(klein15a$variables, khb(X=X, y=Y, z="eta"))
}
\references{
Karlson, K.B., A. Holm and R. Breen (2012). Comparing regression coefficients between same-sample nested models using logit and probit: A new method. \emph{Sociological Methodology}, 42(1):286--313.
}
\author{
Thilo Klein
}
\keyword{summary}
